wip: refactor table components
