--- packages/app-website-builder/src/presentation/redirects/RedirectList/components/Filters/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FilterByStatus.js";
