+++ packages/app-website-builder/src/presentation/redirects/RedirectList/components/getRedirectsLabel.ts
@@ -1,6 +1,3 @@
 export const getRedirectsLabel = (count = 0): string => {
     return `${count} ${count === 1 ? "redirect" : "redirects"}`;
 };
-
-export * from "./BulkActionDelete.js";
-export * from "./BulkActionMove.js";
